Added initial cli test file. Restructured tests directory to be more organized and have more meaningful names
