+++ src/test/java/mike/datastructures/RelationshipTest.java
@@ -1,11 +1,10 @@
-package mike.testcases;
+package mike.datastructures;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertFalse;
 import org.junit.Test;
 
-import mike.datastructures.Relationship;
 import mike.datastructures.Relationship.Type;
 
 public class RelationshipTest {
